+++ app/src/test/java/org/mozilla/fenix/components/TestCore.kt
@@ -9,12 +9,10 @@ import io.mockk.mockk
 import kotlinx.coroutines.ObsoleteCoroutinesApi
 import mozilla.components.browser.engine.gecko.GeckoEngine
 import mozilla.components.browser.session.SessionManager
-import org.mozilla.geckoview.GeckoRuntime
 
 @ObsoleteCoroutinesApi
 class TestCore(private val context: Context) : Core(context) {
 
-    override val runtime = mockk<GeckoRuntime>(relaxed = true)
     override val engine = mockk<GeckoEngine>(relaxed = true)
     override val sessionManager = SessionManager(engine)
 }
